pyproject.toml
[tool.poetry]
name = "pos-backend"
version = "0.1.0"
description = ""
authors = ["bright <brbojr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.10.8"
aiosignal = "1.3.1"
amqp = "5.2.0"
asgiref = "3.8.1"
async-timeout = "4.0.3"
attrs = "24.2.0"
billiard = "4.2.1"
black = "24.8.0"
cachetools = "5.5.0"
celery = "5.4.0"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.3.2"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
click = "8.1.7"
colorama = "0.4.6"
cryptography = "43.0.1"
dj-database-url = "2.2.0"
django-allauth = "65.0.2"
django-cors-headers = "4.4.0"
django-filter = "24.3"
django-storages = "1.14.4"
django = "5.1.1"
djangorestframework-simplejwt = "5.3.1"
djangorestframework = "3.15.2"
flake8 = "7.1.1"
frozenlist = "1.4.1"
gunicorn = "23.0.0"
h11 = "0.14.0"
httplib2 = "0.22.0"
idna = "3.10"
kombu = "5.4.2"
mccabe = "0.7.0"
multidict = "6.1.0"
mypy-extensions = "1.0.0"
oauthlib = "3.2.2"
packaging = "24.1"
pathspec = "0.12.1"
paypalrestsdk = "1.13.3"
pillow = "10.4.0"
platformdirs = "4.3.6"
prompt-toolkit = "3.0.48"
proto-plus = "1.24.0"
protobuf = "5.28.2"
psycopg2-binary = "2.9.9"
pyasn1-modules = "0.4.1"
pyasn1 = "0.6.1"
pycodestyle = "2.12.1"
pycparser = "2.22"
pyflakes = "3.2.0"
pyjwt = "2.9.0"
pyopenssl = "24.2.1"
pyparsing = "3.1.4"
python-dateutil = "2.9.0.post0"
python-decouple = "3.8"
python-dotenv = "1.0.1"
python-http-client = "3.3.7"
redis = "5.1.1"
requests-oauthlib = "2.0.0"
requests = "2.32.3"
rsa = "4.9"
sendgrid = "6.11.0"
six = "1.16.0"
sqlparse = "0.5.1"
starkbank-ecdsa = "2.2.0"
stripe = "11.1.0"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
uvicorn = "0.31.0"
vine = "5.1.0"
wcwidth = "0.2.13"
whitenoise = "6.7.0"
yarl = "1.13.1"
django-heroku = "^0.3.1"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
